# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clash Market API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccountGame = typeof AccountGame[keyof typeof AccountGame];


export const AccountGame = {
  'clash-of-clans': 'clash-of-clans',
  'clash-royale': 'clash-royale',
} as const;

export type AccountStatus = typeof AccountStatus[keyof typeof AccountStatus];


export const AccountStatus = {
  available: 'available',
  reserved: 'reserved',
  sold: 'sold',
} as const;

export interface Account {
  id: number;
  title: string;
  slug: string;
  game: AccountGame;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  images: string[];
  /** @nullable */
  description?: string | null;
  status: AccountStatus;
  /** @nullable */
  townHall?: number | null;
  /** @nullable */
  arena?: string | null;
  /** @nullable */
  trophies?: number | null;
  /** @nullable */
  heroes?: string | null;
  /** @nullable */
  gems?: number | null;
  /** @nullable */
  skins?: string | null;
  /** @nullable */
  league?: string | null;
  /** @nullable */
  evolutions?: string | null;
  /** @nullable */
  emotes?: string | null;
  /** @nullable */
  maxCards?: number | null;
  /** @nullable */
  whatsappMessage?: string | null;
  featured?: boolean;
  createdAt: string;
}

export type AccountInputGame = typeof AccountInputGame[keyof typeof AccountInputGame];


export const AccountInputGame = {
  'clash-of-clans': 'clash-of-clans',
  'clash-royale': 'clash-royale',
} as const;

export type AccountInputStatus = typeof AccountInputStatus[keyof typeof AccountInputStatus];


export const AccountInputStatus = {
  available: 'available',
  reserved: 'reserved',
  sold: 'sold',
} as const;

export interface AccountInput {
  title: string;
  slug: string;
  game: AccountInputGame;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  images?: string[];
  /** @nullable */
  description?: string | null;
  status: AccountInputStatus;
  /** @nullable */
  townHall?: number | null;
  /** @nullable */
  arena?: string | null;
  /** @nullable */
  trophies?: number | null;
  /** @nullable */
  heroes?: string | null;
  /** @nullable */
  gems?: number | null;
  /** @nullable */
  skins?: string | null;
  /** @nullable */
  league?: string | null;
  /** @nullable */
  evolutions?: string | null;
  /** @nullable */
  emotes?: string | null;
  /** @nullable */
  maxCards?: number | null;
  /** @nullable */
  whatsappMessage?: string | null;
  featured?: boolean;
}

export type AccountUpdateGame = typeof AccountUpdateGame[keyof typeof AccountUpdateGame];


export const AccountUpdateGame = {
  'clash-of-clans': 'clash-of-clans',
  'clash-royale': 'clash-royale',
} as const;

export type AccountUpdateStatus = typeof AccountUpdateStatus[keyof typeof AccountUpdateStatus];


export const AccountUpdateStatus = {
  available: 'available',
  reserved: 'reserved',
  sold: 'sold',
} as const;

export interface AccountUpdate {
  title?: string;
  slug?: string;
  game?: AccountUpdateGame;
  price?: number;
  /** @nullable */
  oldPrice?: number | null;
  images?: string[];
  /** @nullable */
  description?: string | null;
  status?: AccountUpdateStatus;
  /** @nullable */
  townHall?: number | null;
  /** @nullable */
  arena?: string | null;
  /** @nullable */
  trophies?: number | null;
  /** @nullable */
  heroes?: string | null;
  /** @nullable */
  gems?: number | null;
  /** @nullable */
  skins?: string | null;
  /** @nullable */
  league?: string | null;
  /** @nullable */
  evolutions?: string | null;
  /** @nullable */
  emotes?: string | null;
  /** @nullable */
  maxCards?: number | null;
  /** @nullable */
  whatsappMessage?: string | null;
  featured?: boolean;
}

export interface MarketplaceStats {
  totalAccounts: number;
  availableAccounts: number;
  cocAccounts: number;
  crAccounts: number;
  featuredCount: number;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  content: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  /** @nullable */
  game?: string | null;
  createdAt: string;
}

export interface BlogPostInput {
  title: string;
  slug: string;
  content: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  /** @nullable */
  game?: string | null;
}

export interface BlogPostUpdate {
  title?: string;
  slug?: string;
  content?: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  /** @nullable */
  game?: string | null;
}

export interface AdminVerifyInput {
  token: string;
}

export interface AdminVerifyResult {
  valid: boolean;
  email: string;
}

export interface ImageUploadInput {
  data: string;
  /** @nullable */
  folder?: string | null;
}

export interface ImageUploadResult {
  url: string;
  publicId: string;
}

export type ListAccountsParams = {
game?: ListAccountsGame;
status?: ListAccountsStatus;
featured?: ListAccountsFeatured;
limit?: number;
offset?: number;
};

export type ListAccountsGame = typeof ListAccountsGame[keyof typeof ListAccountsGame];


export const ListAccountsGame = {
  'clash-of-clans': 'clash-of-clans',
  'clash-royale': 'clash-royale',
} as const;

export type ListAccountsStatus = typeof ListAccountsStatus[keyof typeof ListAccountsStatus];


export const ListAccountsStatus = {
  available: 'available',
  reserved: 'reserved',
  sold: 'sold',
} as const;

export type ListAccountsFeatured = typeof ListAccountsFeatured[keyof typeof ListAccountsFeatured];


export const ListAccountsFeatured = {
  true: 'true',
  false: 'false',
} as const;

export type ListBlogPostsParams = {
game?: string;
limit?: number;
offset?: number;
};