# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clash Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      parameters:
        - name: game
          in: query
          schema:
            type: string
            enum: [clash-of-clans, clash-royale]
        - name: status
          in: query
          schema:
            type: string
            enum: [available, reserved, sold]
        - name: featured
          in: query
          schema:
            type: string
            enum: ["true", "false"]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"

    post:
      operationId: createAccount
      tags: [admin]
      summary: Create account (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
  /accounts/{slug}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found

    put:
      operationId: updateAccount
      tags: [admin]
      summary: Update account (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"

    delete:
      operationId: deleteAccount
      tags: [admin]
      summary: Delete account (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /accounts/{slug}/related:
    get:
      operationId: getRelatedAccounts
      tags: [accounts]
      summary: Get related accounts
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Related accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
...
  /admin/verify:
    post:
      operationId: verifyAdmin
      tags: [admin]
      summary: Verify admin Firebase token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVerifyInput"
      responses:
        "200":
          description: Admin verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVerifyResult"
        "401":
          description: Unauthorized
  /upload/image:
    post:
      operationId: uploadImage
      tags: [admin]
      summary: Upload image to Cloudinary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUploadInput"
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResult"
components:
  schemas:
...
    Account:
      type: object
      required: [id, title, slug, game, price, status, images, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        game:
          type: string
          enum: [clash-of-clans, clash-royale]
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, reserved, sold]
        townHall:
          type: ["integer", "null"]
        arena:
          type: ["string", "null"]
        trophies:
          type: ["integer", "null"]
        heroes:
          type: ["string", "null"]
        gems:
          type: ["integer", "null"]
        skins:
          type: ["string", "null"]
        league:
          type: ["string", "null"]
        evolutions:
          type: ["string", "null"]
        emotes:
          type: ["string", "null"]
        maxCards:
          type: ["integer", "null"]
        whatsappMessage:
          type: ["string", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
    AccountInput:
      type: object
      required: [title, slug, game, price, status]
      properties:
        title:
          type: string
        slug:
          type: string
        game:
          type: string
          enum: [clash-of-clans, clash-royale]
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, reserved, sold]
        townHall:
          type: ["integer", "null"]
        arena:
          type: ["string", "null"]
        trophies:
          type: ["integer", "null"]
        heroes:
          type: ["string", "null"]
        gems:
          type: ["integer", "null"]
        skins:
          type: ["string", "null"]
        league:
          type: ["string", "null"]
        evolutions:
          type: ["string", "null"]
        emotes:
          type: ["string", "null"]
        maxCards:
          type: ["integer", "null"]
        whatsappMessage:
          type: ["string", "null"]
        featured:
          type: boolean
    AccountUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        game:
          type: string
          enum: [clash-of-clans, clash-royale]
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, reserved, sold]
        townHall:
          type: ["integer", "null"]
        arena:
          type: ["string", "null"]
        trophies:
          type: ["integer", "null"]
        heroes:
          type: ["string", "null"]
        gems:
          type: ["integer", "null"]
        skins:
          type: ["string", "null"]
        league:
          type: ["string", "null"]
        evolutions:
          type: ["string", "null"]
        emotes:
          type: ["string", "null"]
        maxCards:
          type: ["integer", "null"]
        whatsappMessage:
          type: ["string", "null"]
        featured:
          type: boolean
...
    AdminVerifyInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
    AdminVerifyResult:
      type: object
      required: [valid, email]
      properties:
        valid:
          type: boolean
        email:
          type: string
    ImageUploadInput:
      type: object
      required: [data]
      properties:
        data:
          type: string
        folder:
          type: ["string", "null"]
    ImageUploadResult:
      type: object
      required: [url, publicId]
      properties:
        url:
          type: string
        publicId:
          type: string