# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clash Market API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountGame } from './accountGame';
import type { AccountStatus } from './accountStatus';

export interface Account {
  id: number;
  title: string;
  slug: string;
  game: AccountGame;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  images: string[];
  /** @nullable */
  description?: string | null;
  status: AccountStatus;
  /** @nullable */
  townHall?: number | null;
  /** @nullable */
  arena?: string | null;
  /** @nullable */
  trophies?: number | null;
  /** @nullable */
  heroes?: string | null;
  /** @nullable */
  gems?: number | null;
  /** @nullable */
  skins?: string | null;
  /** @nullable */
  league?: string | null;
  /** @nullable */
  evolutions?: string | null;
  /** @nullable */
  emotes?: string | null;
  /** @nullable */
  maxCards?: number | null;
  /** @nullable */
  whatsappMessage?: string | null;
  featured?: boolean;
  createdAt: string;
}