# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clash Market API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountInputGame } from './accountInputGame';
import type { AccountInputStatus } from './accountInputStatus';

export interface AccountInput {
  title: string;
  slug: string;
  game: AccountInputGame;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  images?: string[];
  /** @nullable */
  description?: string | null;
  status: AccountInputStatus;
  /** @nullable */
  townHall?: number | null;
  /** @nullable */
  arena?: string | null;
  /** @nullable */
  trophies?: number | null;
  /** @nullable */
  heroes?: string | null;
  /** @nullable */
  gems?: number | null;
  /** @nullable */
  skins?: string | null;
  /** @nullable */
  league?: string | null;
  /** @nullable */
  evolutions?: string | null;
  /** @nullable */
  emotes?: string | null;
  /** @nullable */
  maxCards?: number | null;
  /** @nullable */
  whatsappMessage?: string | null;
  featured?: boolean;
}