# lib/api-zod/src/generated/types/adminVerifyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clash Market API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminVerifyResult {
  valid: boolean;
  email: string;
}